+++ database/migrations/2020_04_04_000004_create_projects_table.php
@@ -30,7 +30,6 @@ class CreateProjectsTable extends Migration
             $table->text('description')->nullable();
             $table->text('notes')->nullable();
             $table->integer('visible')->default(1)->nullable();
-            $table->integer('type')->nullable();
             $table->timestamp('started_at')->nullable();
             $table->timestamp('delivered_at')->nullable();
             $table->timestamp('expected_at')->nullable();