+++ src/Helpers/helpers.php
@@ -1,14 +1,23 @@
 <?php
 
+use Illuminate\Support\Str;
+
 function projects_config_published()
 {
     return file_exists(config_path('projects.php'));
 }
 
+function projects_base_path(string $append = '') : string
+{
+    return Str::replaceLast('/src/Helpers', '', dirname(__FILE__)).$append;
+}
+
 function user_model()
 {
     $user_class = config('projects.user_class');
+
     $repository = (new Illuminate\Config\Repository);
+
     $repository->set('custom', require projects_base_path('/config/projects.php'));
 
     if ( ! $user_class) {
@@ -16,9 +25,4 @@ function user_model()
     }
 
     return new $user_class;
-}
-
-function projects_base_path(string $append = '') : string
-{
-    return \Illuminate\Support\Str::replaceLast('/src/Helpers', '', dirname(__FILE__)).$append;
 }