+++ src/Traits/HasUsers.php
@@ -11,7 +11,7 @@ trait HasUsers
      */
     public function users() : BelongsToMany
     {
-        return $this->belongsToMany(config('projects.user_class'), 'project_users', 'user_id', 'project_id');
+        return $this->belongsToMany(user_model(), 'project_users', 'user_id', 'project_id');
     }
 
     /**