+++ tests/ProjectTest.php
@@ -2,120 +2,125 @@
 
 namespace Chriscreates\Projects\Tests;
 
-use Carbon\Carbon;
 use Chriscreates\Projects\Models\Project;
+use Chriscreates\Projects\Models\Status;
+use Chriscreates\Projects\Models\User;
+use Illuminate\Support\Facades\Hash;
 
 class ProjectTest extends TestCase
 {
-    private $project1;
-
-    public function setUp(): void
+    /** @test */
+    public function it_has_a_title()
     {
-        parent::setUp();
+        $project = Project::create(['title' => 'Title']);
 
-        $this->assertCount(0, Project::all());
-
-        $this->project1 = Project::firstOrCreate(
-            ['title' => 'string1'],
-            [
-                'visible' => true,
-                'started_at' => now()->subMonth(),
-                'delivered_at' => now(),
-                'expected_at' => now(),
-            ]
-        );
+        $this->assertEquals($project->title, 'Title');
     }
 
     /** @test */
-    public function it_can_create_a_project()
+    public function it_has_a_description()
     {
-        $this->assertInstanceOf(Project::class, $this->project1);
+        $project = Project::create([
+            'title' => 'Title',
+            'description' => 'This is a description',
+        ]);
 
-        $this->assertCount(1, Project::all());
-
-        $this->assertSame('string1', $this->project1->title);
+        $this->assertEquals($project->description, 'This is a description');
     }
 
     /** @test */
-    public function it_has_visibility()
+    public function it_has_a_notes()
     {
-        $project2 = Project::firstOrCreate(
-            ['title' => 'string2'],
-            [
-                'visible' => 0,
-            ]
-        );
+        $project = Project::create([
+            'title' => 'Title',
+            'notes' => 'These are some notes',
+        ]);
 
-        $project3 = Project::firstOrCreate(
-            ['title' => 'string3'],
-            [
-                'visible' => false,
-            ]
-        );
+        $this->assertEquals($project->notes, 'These are some notes');
+    }
 
-        $this->assertTrue($this->project1->isVisible());
+    /** @test */
+    public function it_has_togglable_visiblitity()
+    {
+        $project = Project::create([
+            'title' => 'Title',
+            'visible' => true,
+        ]);
 
-        $this->assertTrue($this->project1->visible);
+        $this->assertSame($project->visible, true);
+    }
 
-        $this->assertFalse($project2->isVisible());
-
-        $this->assertTrue($project3->isNotVisible());
+    /** @test */
+    public function it_has_a_started_at_date()
+    {
+        $project = Project::create([
+            'title' => 'Title',
+            'started_at' => now(),
+        ]);
 
         $this->assertEquals(
-            ['string1', 'string2', 'string3'],
-            Project::pluck('title')->toArray()
+            $project->started_at->toDateTimeString(),
+            now()->toDateTimeString()
         );
-
-        $this->assertCount(3, Project::all());
     }
 
     /** @test */
-    public function it_has_a_due_date()
+    public function it_has_a_delivered_at_date()
     {
-        // Days
-        $days = $this->project1->dueIn('days');
-        $last_month_days = Carbon::now()->subMonth(1)->daysInMonth;
-
-        $this->assertEquals($days, $last_month_days);
-        $this->assertSame($days, $last_month_days);
-        $this->assertTrue($days == $last_month_days);
-
-        // Months
-        $project2 = Project::firstOrCreate(
-            ['title' => 'string2'],
-            [
-                'started_at' => now()->subMonths(4),
-                'expected_at' => now()->addMonth(1),
-            ]
-        );
-
-        $months = $project2->dueIn('months');
-        $this->assertEquals($months, 5);
-        $this->assertSame($months, 5);
-        $this->assertTrue($months == 5);
-
-        // Years
-        $project3 = Project::firstOrCreate(
-            ['title' => 'string3'],
-            [
-                'started_at' => now()->subYear(1),
-                'expected_at' => now()->addMonths(2),
-            ]
-        );
-
-        $years = $project3->dueIn('years');
-        $this->assertEquals($years, 1);
-        $this->assertSame($years, 1);
-        $this->assertTrue($years == 1);
+        $project = Project::create([
+            'title' => 'Title',
+            'delivered_at' => now(),
+        ]);
 
         $this->assertEquals(
-            ['string1', 'string2', 'string3'],
-            Project::pluck('title')->toArray()
+            $project->delivered_at->toDateTimeString(),
+            now()->toDateTimeString()
         );
-
-        $this->assertCount(3, Project::all());
     }
 
-    // TODO
-    // Has Comments
+    /** @test */
+    public function it_has_a_expected_at_date()
+    {
+        $project = Project::create([
+            'title' => 'Title',
+            'expected_at' => now(),
+        ]);
+
+        $this->assertEquals(
+            $project->expected_at->toDateTimeString(),
+            now()->toDateTimeString()
+        );
+    }
+
+    /** @test */
+    public function it_has_can_be_associated_to_a_user()
+    {
+        $project = Project::create(['title' => 'Title']);
+
+        $user = User::create([
+            'name' => 'Test',
+            'email' => 'test@test.com',
+            'password' => Hash::make('password'),
+        ]);
+
+        $this->assertNull($project->owner);
+
+        $project->owner()->associate($user);
+
+        $this->assertInstanceOf(User::class, $project->owner);
+    }
+
+    /** @test */
+    public function it_has_can_be_associated_to_a_status()
+    {
+        $project = Project::create(['title' => 'Title']);
+
+        $status = Status::create(['name' => 'Test']);
+
+        $this->assertNull($project->status);
+
+        $project->status()->associate($status);
+
+        $this->assertInstanceOf(Status::class, $project->status);
+    }
 }