+++ tests/TestCase.php
@@ -3,14 +3,11 @@
 namespace Chriscreates\Projects\Tests;
 
 use Chriscreates\Projects\ProjectsServiceProvider;
-// use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Facades\Schema;
 use Orchestra\Testbench\TestCase as Orchestra;
 
 abstract class TestCase extends Orchestra
 {
-    // use RefreshDatabase;
-
     public function setUp(): void
     {
         parent::setUp();
@@ -33,6 +30,7 @@ abstract class TestCase extends Orchestra
         Schema::dropAllTables();
 
         $this->loadMigrationsFrom(projects_base_path('/database/test_migrations'));
+
         $this->loadMigrationsFrom(projects_base_path('/database/migrations'));
     }
 }